ENGR00259949-1 mxc vout:do not dqueue all bufs before streamoff
In ic bypass case, we cannot get a buffer dqueued before the
next 2 buffers are put on display channel, which may block
us from dequeuing the last 2 buffers at the end of the video
clip. However, we don't have to dqueue all buffers before
streaming off in the unit test. This patch changes to stream
off directly, but not to wait until all buffers are dequeued.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>