ENGR00291326 mx6: Add USB camera unit test
This patch adds the test support for the USB camera (UVC driver).
Since the UVC driver does not implement some IOCTLs like DBG_G_CHIP_IDENT,
S_CROP and G_CROP, the patch adds the judgement for them if USB camera
is used. This patch should not impact the existing features for mxc capture
test program.

The newly-added command option is
-uvc:  needed for USB camera test

The command is as follows.

./mxc_v4l2_capture.out -uvc -f YUYV -d /dev/video1 -ow 320 -oh 240 -c 100 1.yuv

Signed-off-by: Robby Cai <R63905@freescale.com>